refactor: #8 index.ts 에 import/export 문 넣어놓기
