+++ src/components/index.ts
@@ -2,5 +2,22 @@ import Header from './Header/Header';
 import ProgressBar from './meeteam/progressBar/ProgressBar';
 import Menu from './meeteam/menu/Menu';
 import Main from './meeteam/main/Main';
+import Information from './meeteam/main/information/Information';
+import Member from './meeteam/main/member/Member';
+import MemberProfile from './meeteam/main/member/MemberProfile';
+import MemberRadioBox from './meeteam/main/member/MemberRadioBox';
+import WorkRegistration from './meeteam/main/workRegistration/WorkRegistration';
+import Setting from './meeteam/main/setting/Setting';
 
-export { Header, ProgressBar, Menu, Main };
+export {
+	Header,
+	ProgressBar,
+	Menu,
+	Main,
+	Information,
+	Member,
+	MemberProfile,
+	MemberRadioBox,
+	WorkRegistration,
+	Setting,
+};
